move historical data store
