+++ packages/web/components/complex/add-conc-liquidity.tsx
@@ -35,9 +35,11 @@ import {
   useAmplitudeAnalytics,
   useTranslation,
 } from "~/hooks";
-import { useHistoricalAndLiquidityData } from "~/hooks/ui-config/use-historical-and-depth-data";
+import {
+  ObservableHistoricalAndLiquidityData,
+  useHistoricalAndLiquidityData,
+} from "~/hooks/ui-config/use-historical-and-depth-data";
 import { useStore } from "~/stores";
-import { ObservableHistoricalAndLiquidityData } from "~/stores/derived-data";
 import { formatPretty } from "~/utils/formatter";
 
 import { Tooltip } from "../tooltip";
